Move configuration handling into its own module
