+++ mqtt_gateway.py
@@ -1,4 +1,3 @@
-import argparse
 import asyncio
 import datetime
 import faulthandler
@@ -8,8 +7,7 @@ import os
 import signal
 import sys
 import time
-import urllib.parse
-from typing import Callable, override
+from typing import override
 
 import apscheduler.schedulers.asyncio
 from saic_ismart_client_ng import SaicApi
@@ -23,7 +21,8 @@ from saic_ismart_client_ng.exceptions import SaicApiException
 from saic_ismart_client_ng.model import SaicApiConfiguration
 
 import mqtt_topics
-from configuration import Configuration, TransportProtocol
+from configuration import Configuration
+from configuration.parser import process_arguments
 from exceptions import MqttGatewayException
 from integrations.abrp.api import AbrpApi, AbrpApiException
 from integrations.home_assistant.discovery import HomeAssistantDiscovery
@@ -34,7 +33,6 @@ from saic_api_listener import MqttGatewaySaicApiListener, MqttGatewayAbrpListene
 from vehicle import RefreshMode, VehicleState
 
 MSG_CMD_SUCCESSFUL = 'Success'
-CHARGING_STATIONS_FILE = 'charging-stations.json'
 
 
 def epoch_value_to_str(time_value: int) -> str:
@@ -690,268 +688,6 @@ class MessageHandler:
             return False
         else:
             return True
-
-
-class EnvDefault(argparse.Action):
-    def __init__(self, envvar, required=True, default=None, **kwargs):
-        if (
-                envvar in os.environ
-                and os.environ[envvar]
-        ):
-            default = os.environ[envvar]
-        if required and default:
-            required = False
-        super(EnvDefault, self).__init__(default=default, required=required, **kwargs)
-
-    def __call__(self, parser, namespace, values, option_string=None):
-        setattr(namespace, self.dest, values)
-
-
-def process_arguments() -> Configuration:
-    config = Configuration()
-    parser = argparse.ArgumentParser(prog='MQTT Gateway')
-    try:
-        parser.add_argument('-m', '--mqtt-uri',
-                            help='The URI to the MQTT Server. Environment Variable: MQTT_URI,'
-                                 + 'TCP: tcp://mqtt.eclipseprojects.io:1883 '
-                                 + 'WebSocket: ws://mqtt.eclipseprojects.io:9001'
-                                 + 'TLS: tls://mqtt.eclipseprojects.io:8883',
-                            dest='mqtt_uri', required=True, action=EnvDefault, envvar='MQTT_URI')
-        parser.add_argument('--mqtt-server-cert',
-                            help='Path to the server certificate authority file in PEM format for TLS.',
-                            dest='tls_server_cert_path', required=False, action=EnvDefault, envvar='MQTT_SERVER_CERT')
-        parser.add_argument('--mqtt-user', help='The MQTT user name. Environment Variable: MQTT_USER',
-                            dest='mqtt_user', required=False, action=EnvDefault, envvar='MQTT_USER')
-        parser.add_argument('--mqtt-password',
-                            help='The MQTT password. Environment Variable: MQTT_PASSWORD', dest='mqtt_password',
-                            required=False, action=EnvDefault, envvar='MQTT_PASSWORD')
-        parser.add_argument('--mqtt-client-id', help='The MQTT Client Identifier. Environment Variable: '
-                                                     + 'MQTT_CLIENT_ID '
-                                                     + 'Default is saic-python-mqtt-gateway',
-                            default='saic-python-mqtt-gateway', dest='mqtt_client_id', required=False,
-                            action=EnvDefault, envvar='MQTT_CLIENT_ID')
-        parser.add_argument('--mqtt-topic-prefix',
-                            help='MQTT topic prefix. Environment Variable: MQTT_TOPIC Default is saic', default='saic',
-                            dest='mqtt_topic', required=False, action=EnvDefault, envvar='MQTT_TOPIC')
-        parser.add_argument('-s', '--saic-rest-uri',
-                            help='The SAIC uri. Environment Variable: SAIC_REST_URI Default is the European '
-                                 'Production Endpoint: https://tap-eu.soimt.com',
-                            default='https://gateway-mg-eu.soimt.com/api.app/v1/', dest='saic_rest_uri', required=False,
-                            action=EnvDefault,
-                            envvar='SAIC_REST_URI')
-        parser.add_argument('-u', '--saic-user',
-                            help='The SAIC user name. Environment Variable: SAIC_USER', dest='saic_user', required=True,
-                            action=EnvDefault, envvar='SAIC_USER')
-        parser.add_argument('-p', '--saic-password',
-                            help='The SAIC password. Environment Variable: SAIC_PASSWORD', dest='saic_password',
-                            required=True, action=EnvDefault, envvar='SAIC_PASSWORD')
-        parser.add_argument('--saic-phone-country-code',
-                            help='The SAIC phone country code. Environment Variable: SAIC_PHONE_COUNTRY_CODE',
-                            dest='saic_phone_country_code', required=False, action=EnvDefault,
-                            envvar='SAIC_PHONE_COUNTRY_CODE')
-        parser.add_argument('--saic-region', '--saic-region',
-                            help='The SAIC API region. Environment Variable: SAIC_REGION', default='eu',
-                            dest='saic_region', required=False, action=EnvDefault, envvar='SAIC_REGION')
-        parser.add_argument('--saic-tenant-id',
-                            help='The SAIC API tenant id. Environment Variable: SAIC_TENANT_ID', default='459771',
-                            dest='saic_tenant_id', required=False, action=EnvDefault,
-                            envvar='SAIC_TENANT_ID')
-        parser.add_argument('--abrp-api-key',
-                            help='The API key for the A Better Route Planer telemetry API.'
-                                 + ' Default is the open source telemetry'
-                                 + ' API key 8cfc314b-03cd-4efe-ab7d-4431cd8f2e2d.'
-                                 + ' Environment Variable: ABRP_API_KEY',
-                            default='8cfc314b-03cd-4efe-ab7d-4431cd8f2e2d', dest='abrp_api_key', required=False,
-                            action=EnvDefault, envvar='ABRP_API_KEY')
-        parser.add_argument('--abrp-user-token', help='The mapping of VIN to ABRP User Token.'
-                                                      + ' Multiple mappings can be provided seperated by ,'
-                                                      + ' Example: LSJXXXX=12345-abcdef,LSJYYYY=67890-ghijkl,'
-                                                      + ' Environment Variable: ABRP_USER_TOKEN',
-                            dest='abrp_user_token', required=False, action=EnvDefault, envvar='ABRP_USER_TOKEN')
-        parser.add_argument('--battery-capacity-mapping', help='The mapping of VIN to full batteryc'
-                                                               + ' apacity. Multiple mappings can be provided separated'
-                                                               + ' by , Example: LSJXXXX=54.0,LSJYYYY=64.0,'
-                                                               + ' Environment Variable: BATTERY_CAPACITY_MAPPING',
-                            dest='battery_capacity_mapping', required=False, action=EnvDefault,
-                            envvar='BATTERY_CAPACITY_MAPPING')
-        parser.add_argument('--charging-stations-json',
-                            help='Custom charging stations configuration file name', dest='charging_stations_file',
-                            required=False, action=EnvDefault, envvar='CHARGING_STATIONS_JSON')
-        parser.add_argument('--saic-relogin-delay',
-                            help='How long to wait before attempting another login to the SAIC API. Environment '
-                                 'Variable: SAIC_RELOGIN_DELAY', dest='saic_relogin_delay', required=False,
-                            action=EnvDefault, envvar='SAIC_RELOGIN_DELAY', type=check_positive)
-        parser.add_argument('--ha-discovery',
-                            help='Enable Home Assistant Discovery. Environment Variable: HA_DISCOVERY_ENABLED',
-                            dest='ha_discovery_enabled', required=False,
-                            action=EnvDefault,
-                            envvar='HA_DISCOVERY_ENABLED', default=True, type=check_bool)
-        parser.add_argument('--ha-discovery-prefix',
-                            help='Home Assistant Discovery Prefix. Environment Variable: HA_DISCOVERY_PREFIX',
-                            dest='ha_discovery_prefix', required=False, action=EnvDefault, envvar='HA_DISCOVERY_PREFIX',
-                            default='homeassistant')
-        parser.add_argument('--ha-show-unavailable',
-                            help='Show entities as Unavailable in Home Assistant when car polling fails. '
-                                 'Environment Variable: HA_SHOW_UNAVAILABLE', dest='ha_show_unavailable',
-                            required=False, action=EnvDefault, envvar='HA_SHOW_UNAVAILABLE', default=True,
-                            type=check_bool)
-        parser.add_argument('--messages-request-interval',
-                            help='The interval for retrieving messages in seconds. Environment Variable: '
-                                 'MESSAGES_REQUEST_INTERVAL', dest='messages_request_interval',
-                            required=False, action=EnvDefault,
-                            envvar='MESSAGES_REQUEST_INTERVAL', default=60)
-        parser.add_argument('--charge-min-percentage',
-                            help='How many % points we should try to refresh the charge state. Environment Variable: '
-                                 'CHARGE_MIN_PERCENTAGE', dest='charge_dynamic_polling_min_percentage', required=False,
-                            action=EnvDefault, envvar='CHARGE_MIN_PERCENTAGE', default='1.0', type=check_positive_float)
-        parser.add_argument('--publish-raw-api-data',
-                            help='Publish raw SAIC API request/response to MQTT. Environment Variable: '
-                                 'PUBLISH_RAW_API_DATA_ENABLED',
-                            dest='publish_raw_api_data', required=False,
-                            action=EnvDefault,
-                            envvar='PUBLISH_RAW_API_DATA_ENABLED', default=False, type=check_bool)
-        parser.add_argument('--publish-raw-abrp-data',
-                            help='Publish raw ABRP API request/response to MQTT. Environment Variable: '
-                                 'PUBLISH_RAW_ABRP_DATA_ENABLED',
-                            dest='publish_raw_abrp_data', required=False, action=EnvDefault,
-                            envvar='PUBLISH_RAW_ABRP_DATA_ENABLED', default=False, type=check_bool)
-
-        args = parser.parse_args()
-        config.mqtt_user = args.mqtt_user
-        config.mqtt_password = args.mqtt_password
-        config.mqtt_client_id = args.mqtt_client_id
-        config.charge_dynamic_polling_min_percentage = args.charge_dynamic_polling_min_percentage
-        if args.saic_relogin_delay:
-            config.saic_relogin_delay = args.saic_relogin_delay
-        config.mqtt_topic = args.mqtt_topic
-        config.saic_rest_uri = args.saic_rest_uri
-        config.saic_region = args.saic_region
-        config.saic_tenant_id = str(args.saic_tenant_id)
-        config.saic_user = args.saic_user
-        config.saic_password = args.saic_password
-        config.saic_phone_country_code = args.saic_phone_country_code
-        config.abrp_api_key = args.abrp_api_key
-        if args.abrp_user_token:
-            cfg_value_to_dict(args.abrp_user_token, config.abrp_token_map)
-        if args.battery_capacity_mapping:
-            cfg_value_to_dict(
-                args.battery_capacity_mapping,
-                config.battery_capacity_map,
-                value_type=check_positive_float
-            )
-        if args.charging_stations_file:
-            process_charging_stations_file(config, args.charging_stations_file)
-        else:
-            process_charging_stations_file(config, f'./{CHARGING_STATIONS_FILE}')
-
-        config.saic_password = args.saic_password
-
-        if args.ha_discovery_enabled is not None:
-            config.ha_discovery_enabled = args.ha_discovery_enabled
-
-        if args.publish_raw_api_data is not None:
-            config.publish_raw_api_data = args.publish_raw_api_data
-
-        if args.publish_raw_abrp_data is not None:
-            config.publish_raw_abrp_data = args.publish_raw_abrp_data
-
-        if args.ha_show_unavailable is not None:
-            config.ha_show_unavailable = args.ha_show_unavailable
-
-        if args.ha_discovery_prefix:
-            config.ha_discovery_prefix = args.ha_discovery_prefix
-
-        try:
-            config.messages_request_interval = int(args.messages_request_interval)
-        except ValueError:
-            raise SystemExit(f'No valid integer value for messages_request_interval: {args.messages_request_interval}')
-
-        parse_result = urllib.parse.urlparse(args.mqtt_uri)
-        if parse_result.scheme == 'tcp':
-            config.mqtt_transport_protocol = TransportProtocol.TCP
-        elif parse_result.scheme == 'ws':
-            config.mqtt_transport_protocol = TransportProtocol.WS
-        elif parse_result.scheme == 'tls':
-            config.mqtt_transport_protocol = TransportProtocol.TLS
-            if args.tls_server_cert_path:
-                config.tls_server_cert_path = args.tls_server_cert_path
-            else:
-                raise SystemExit(f'No server certificate authority file provided for TLS MQTT URI {args.mqtt_uri}')
-        else:
-            raise SystemExit(f'Invalid MQTT URI scheme: {parse_result.scheme}, use tcp or ws')
-
-        if not parse_result.port:
-            if config.mqtt_transport_protocol == 'tcp':
-                config.mqtt_port = 1883
-            else:
-                config.mqtt_port = 9001
-        else:
-            config.mqtt_port = parse_result.port
-
-        config.mqtt_host = str(parse_result.hostname)
-
-        return config
-    except argparse.ArgumentError as err:
-        parser.print_help()
-        SystemExit(err)
-
-
-def process_charging_stations_file(config: Configuration, json_file: str):
-    try:
-        with open(json_file, 'r') as f:
-            data = json.load(f)
-
-            for item in data:
-                charge_state_topic = item['chargeStateTopic']
-                charging_value = item['chargingValue']
-                vin = item['vin']
-                if 'socTopic' in item:
-                    charging_station = ChargingStation(vin, charge_state_topic, charging_value, item['socTopic'])
-                else:
-                    charging_station = ChargingStation(vin, charge_state_topic, charging_value)
-                if 'rangeTopic' in item:
-                    charging_station.range_topic = item['rangeTopic']
-                if 'chargerConnectedTopic' in item:
-                    charging_station.connected_topic = item['chargerConnectedTopic']
-                if 'chargerConnectedValue' in item:
-                    charging_station.connected_value = item['chargerConnectedValue']
-                config.charging_stations_by_vin[vin] = charging_station
-    except FileNotFoundError:
-        LOG.warning(f'File {json_file} does not exist')
-    except json.JSONDecodeError as e:
-        LOG.exception(f'Reading {json_file} failed', exc_info=e)
-
-
-def cfg_value_to_dict(cfg_value: str, result_map: dict, value_type: Callable[[str], any] = str):
-    if ',' in cfg_value:
-        map_entries = cfg_value.split(',')
-    else:
-        map_entries = [cfg_value]
-
-    for entry in map_entries:
-        if '=' in entry:
-            key_value_pair = entry.split('=')
-            key = key_value_pair[0]
-            value = key_value_pair[1]
-            result_map[key] = value_type(value)
-
-
-def check_positive(value):
-    ivalue = int(value)
-    if ivalue <= 0:
-        raise argparse.ArgumentTypeError(f'{ivalue} is an invalid positive int value')
-    return ivalue
-
-
-def check_positive_float(value):
-    fvalue = float(value)
-    if fvalue <= 0:
-        raise argparse.ArgumentTypeError(f'{fvalue} is an invalid positive float value')
-    return fvalue
-
-
-def check_bool(value):
-    return str(value).lower() in ['true', '1', 'yes', 'y']
 
 
 if __name__ == '__main__':
